moved localization config to static function of UIManager
